requirements.txt
numpy>=1.21.2
python-telegram-bot>=12.8